Create Shiny
Title: I don't want **columnnames / rownames to be repeated** in RpivotTable output when exported to csv/excel?
When i download RpivotTable output to csv/excel, i get repetition of labels (rownames / columnnames) automatically which are wrongly printed, I don't require the repetition of labels.
I want the exported data to be shown as it is in RpivotTable. I have tried a lot to work it out but all in vain. I would really appreciate if someone could guide me. Thanks